fix: retrieve RepoDataRecord from Rust instead of libsolv
