--- crates/rattler_solve/src/libsolv/custom_keys.rs
@@ -1,1 +1,0 @@
-pub const SOLVABLE_REAL_URL: &str = "solvable:real_repo_url";
