+++ crates/rattler_solve/src/libsolv/wrapper/ffi.rs
