+++ crates/rattler_solve/src/libsolv/wrapper/flags.rs
@@ -1,4 +1,4 @@
-use crate::libsolv::ffi::{
+use super::ffi::{
     SOLVER_ERASE, SOLVER_FLAG_ALLOW_DOWNGRADE, SOLVER_FLAG_ALLOW_UNINSTALL, SOLVER_INSTALL,
     SOLVER_SOLVABLE_PROVIDES, SOLVER_UPDATE,
 };
