+++ crates/rattler_solve/src/libsolv/wrapper/pool.rs
