+++ crates/rattler_solve/src/libsolv/wrapper/queue.rs
