+++ crates/rattler_solve/src/libsolv/wrapper/solver.rs
